Strong typing enabled for CLIPMODE, DEVICE, CCF, PERMIT, KQ

+++ src/fluid/fluid_input.cpp
@@ -434,7 +434,7 @@ static void key_event(struct finput *Input, evKey *Event, LONG Size)
    lua_rawgeti(prv->Lua, LUA_REGISTRYINDEX, Input->Callback); // Get the function reference in Lua and place it on the stack
    lua_rawgeti(prv->Lua, LUA_REGISTRYINDEX, Input->InputValue); // Arg: Input value registered by the client
    lua_pushinteger(prv->Lua, Input->SurfaceID);  // Arg: Surface (if applicable)
-   lua_pushinteger(prv->Lua, Event->Qualifiers); // Arg: Key Flags
+   lua_pushinteger(prv->Lua, ULONG(Event->Qualifiers)); // Arg: Key Flags
    lua_pushinteger(prv->Lua, Event->Code);       // Arg: Key Value
    lua_pushinteger(prv->Lua, Event->Unicode);    // Arg: Unicode character
